Refactor into own module, more functionality, instances
