+++ CHANGELOG.md
@@ -2,19 +2,23 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
-## 3.14.2 -- 2022-11-04
+## 3.14.2 -- 2022-11-08
 
 ### Added
 
 * `AssetClass` now has an `UnsafeFromData` instance, derived the same way as
   `FromData` and `ToData`.
-* `AnyTokenAssetClass`, designating an `AssetClass` whose `TokenName` is
-  irrelevant.
-* `FixedTokenAssetClass`, designating an `AssetClass` whose `TokenName` is
-  significant.
-* `ExtendedAssetClass`, designed to provide a runtime distinction between 
-  `AssetClass`es whose `TokenNames` are arbitrary versus non-arbitrary.
-
+* Module `Plutarch.Extra.ExtendedAssetClass`:
+  * `AnyTokenAssetClass`, designating an `AssetClass` whose `TokenName` is
+    irrelevant.
+  * `FixedTokenAssetClass`, designating an `AssetClass` whose `TokenName` is
+    significant.
+  * `ExtendedAssetClass`, designed to provide a runtime distinction between 
+    `AssetClass`es whose `TokenNames` are arbitrary versus non-arbitrary.
+  * Plutarch equivalents to the above: `PAnyTokenAssetClass`,
+    `PFixedTokenAssetClass` and `PExtendedAssetClass.
+  * Helper functions for comparing and retrieving values from
+    `PExtendedAssetClass`.
 
 ## 3.14.1 -- 2022-11-02
