+++ liqwid-plutarch-extra.cabal
@@ -100,6 +100,7 @@ library
     Plutarch.Extra.DebuggableScript
     Plutarch.Extra.Deriving
     Plutarch.Extra.ExchangeRate
+    Plutarch.Extra.ExtendedAssetClass
     Plutarch.Extra.Field
     Plutarch.Extra.Fixed
     Plutarch.Extra.FixedDecimal
